requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==2.2.6

# Testing
mypy==0.650
pytest==5.4.1
pytest-sugar==0.9.3
pytest-django==3.9.0
factory-boy==2.12.0

# Code quality
flake8==3.6.0
